Add unit test, split executePoolRebalanceLeaf tests in two files to take advantage of parallel test runs in CI
